requirements.txt
# Core dependencies
python-dotenv==1.0.0
openai==1.12.0
beautifulsoup4==4.12.3
playwright==1.42.0
duckduckgo-search==4.4.3
aiohttp==3.9.3
reportlab==4.1.0

# Async and concurrency
asyncio==3.4.3
typing-extensions==4.9.0

# Date and time handling
python-dateutil==2.8.2

# Logging and utilities
colorama==0.4.6
tqdm==4.66.2

# Optional but recommended for better performance
ujson==5.9.0
aiofiles==23.2.1
